requirements.txt
attrs==19.3.0
autopep8==1.5.1
click==7.1.1
dnspython==1.16.0
eventlet==0.25.2
Flask==1.1.2
Flask-SocketIO==4.2.1
greenlet==0.4.15
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
monotonic==1.5
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
py==1.8.1
pycodestyle==2.5.0
pyparsing==2.4.7
pytest==5.4.1
python-dotenv==0.13.0
python-engineio==3.12.1
python-socketio==4.5.1
redis==3.4.1
six==1.14.0
wcwidth==0.1.9
Werkzeug==1.0.1
